refactor: Extract Google ID token verification logic into `GoogleTokenServiceImpl` and simplify application configuration
