+++ src/main/java/com/yaquodorg/yaquod/config/ApplicationConfig.java
@@ -1,8 +1,16 @@
 package com.yaquodorg.yaquod.config;
 
-import java.io.IOException;
-import java.io.InputStream;
-
+import com.google.api.client.googleapis.auth.oauth2.GoogleIdTokenVerifier;
+import com.google.api.client.http.javanet.NetHttpTransport;
+import com.google.api.client.json.gson.GsonFactory;
+import com.google.auth.oauth2.GoogleCredentials;
+import com.google.firebase.FirebaseApp;
+import com.google.firebase.FirebaseOptions;
+import com.google.firebase.messaging.FirebaseMessaging;
+import com.yaquodorg.yaquod.service.user.UserService;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.io.ClassPathResource;
@@ -12,19 +20,22 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 
-import com.google.auth.oauth2.GoogleCredentials;
-import com.google.firebase.FirebaseApp;
-import com.google.firebase.FirebaseOptions;
-import com.google.firebase.messaging.FirebaseMessaging;
-import com.yaquodorg.yaquod.service.user.UserService;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
+import java.io.IOException;
+import java.io.InputStream;
 
 @Configuration
 @RequiredArgsConstructor
 @Slf4j
 public class ApplicationConfig {
+    @Value("${spring.security.oauth2.client.registration.google.client-id}")
+    private String webClientId;
+
+    @Value("${google.android.client-id}")
+    private String androidClientId;
+
+    @Value("${google.ios.client-id}")
+    private String iosClientId;
+
     private final UserService userService;
 
     @Bean
@@ -77,5 +88,35 @@ public class ApplicationConfig {
             log.error("Failed to initialize Firebase", e);
             throw e;
         }
+    }
+
+
+    /**
+     * Builds the GoogleIdTokenVerifier with all possible client IDs
+     * (Web, Android, iOS) to support tokens from different platforms.
+     */
+    @Bean
+    public GoogleIdTokenVerifier buildVerifier() {
+        log.debug("Building Google ID token verifier");
+
+        // Build list of valid client IDs (web, android, ios)
+        java.util.List<String> clientIds = new java.util.ArrayList<>();
+        clientIds.add(webClientId);
+
+        if (androidClientId != null && !androidClientId.isEmpty()) {
+            clientIds.add(androidClientId);
+            log.debug("Android client ID configured");
+        }
+
+        if (iosClientId != null && !iosClientId.isEmpty()) {
+            clientIds.add(iosClientId);
+            log.debug("iOS client ID configured");
+        }
+
+        log.debug("Verifier configured with {} client ID(s)", clientIds.size());
+
+        return new GoogleIdTokenVerifier.Builder(new NetHttpTransport(), GsonFactory.getDefaultInstance())
+                .setAudience(clientIds)
+                .build();
     }
 }
